+++ seleniumide/src/main/java/org/musetest/seleniumide/StepConverter.java
@@ -8,6 +8,5 @@ import org.musetest.core.step.*;
 public interface StepConverter
     {
     StepConfiguration convertStep(TestConverter converter, String command, String param1, String param2) throws UnsupportedError;
-    String[] getCommands();
     }